+++ packages/components/pages/f-account-info/src/components/_tests/AccountInfo.test.js
@@ -1,3 +1,4 @@
+import Vuex from 'vuex';
 import { shallowMount } from '@vue/test-utils';
 import AccountInfo from '../AccountInfo.vue';
 import {
@@ -7,11 +8,21 @@ import {
     token
 } from '../../../test-utils/setup';
 
+const createStore = ({ state = {}, actions = {} } = {}) => new Vuex.Store({
+    modules: {
+        fAccountInfoModule: {
+            state,
+            actions,
+            namespaced: true
+        }
+    }
+});
+
 let sutProps;
 
 describe('AccountInfo', () => {
     beforeEach(() => {
-        // Arrange & Act
+        // Arrange
         sutProps = {
             authToken: token,
             isAuthFinished: true,
@@ -29,7 +40,8 @@ describe('AccountInfo', () => {
             shallowMount(AccountInfo, {
                 i18n,
                 localVue,
-                propsData: { ...sutProps, isAuthFinished: false }
+                propsData: { ...sutProps, isAuthFinished: false },
+                store: createStore()
             });
 
             // Assert
@@ -45,7 +57,8 @@ describe('AccountInfo', () => {
             const wrapper = shallowMount(AccountInfo, {
                 i18n,
                 localVue,
-                propsData: { ...sutProps, isAuthFinished: false }
+                propsData: { ...sutProps, isAuthFinished: false },
+                store: createStore()
             });
 
             // Assert