+++ packages/tools/generator-component/generators/app/templates/test/accessibility/axe-accessibility.spec.js
@@ -1,6 +1,6 @@
 import { getAccessibilityTestResults } from '../../../../../../test/utils/axe-helper';
 
-const <%= name.filename %> = require('../../test-utils/component-objects/f-<%= name.class %>.component');
+const <%= name.filename %> = require('../../test-utils/component-objects/f-<%= name.default %>.component');
 
 const <%= name.class %> = new <%= name.filename %>();
 
@@ -9,9 +9,9 @@ describe('Accessibility tests', () => {
         <%= name.class %>.load();
     });
 
-    it('a11y - should test f-<%= name.class %> component WCAG compliance', () => {
+    it('a11y - should test f-<%= name.default %> component WCAG compliance', () => {
         // Act
-        const axeResults = getAccessibilityTestResults('f-<%= name.class %>');
+        const axeResults = getAccessibilityTestResults('f-<%= name.default %>');
 
         // Assert
         expect(axeResults.violations.length).toBe(0);