+++ packages/tools/generator-component/generators/app/templates/test/component/f-skeleton.component.__spec__.js
@@ -1,13 +1,13 @@
-const <%= name.filename %> = require('../../test-utils/component-objects/f-<%= name.class %>.component');
+const <%= name.filename %> = require('../../test-utils/component-objects/f-<%= name.default %>.component');
 
 const <%= name.class %> = new <%= name.filename %>();
 
-describe('f-<%= name.class %> component tests', () => {
+describe('f-<%= name.default %> component tests', () => {
     beforeEach(() => {
         <%= name.class %>.load();
     });
 
-    it('should display the f-<%= name.class %> component', () => {
+    it('should display the f-<%= name.default %> component', () => {
         // Assert
         expect(<%= name.class %>.isComponentDisplayed()).toBe(true);
     });