+++ packages/components/pages/f-account-info/CHANGELOG.md
@@ -9,7 +9,7 @@ v0.9.0
 *December 08, 2021*
 
 ### Changed
-- Wired up the new Vuex Store that exposes the State `consumer` to the Template plus wired up the load function
+- Wired up the new Vuex Store that exposes the State `consumer` to the Template plus wired up the load action
 
 
 v0.8.0